warning-disclaimer, removed phrase & rephrased beare-metal
changed the warning disclaimer to use the same text as in first-steps (except for the hint at go further, as it's use is too obvious)

empty phrase "We will offer you a number of practical tips." removed. It's a guide. you expect practical tips in the first place...

Windows can be installed on bare metal and does not have iptables. => changed bare-metal to "Linux and UNIX"