clarified some wordings
It is **advised** to have a drop-all rule in 19 for the firewall to actually filter unwanted traffic
the previous wording implied, that it's already the automatic default

The rules are sorted (but not by timestamp or antything. rather by assigned priority. so i removed "chronological")

It's not OVH that blocks port25 traffic in the example; It's the firewall.