Update guide.en-gb.md : corrected example query
Correction in the en-gb guide of "first steps api" : /domains/ does not exist in the OVH API. In a previous commit, I corrected the last query URL with /domain. However, some additional changes were needed to keep the guide consistent.
In this commit : all "/domains/" in queries were replaced with "/domain/" , and the signature hash was re-computed to fit with the change.