details about port to let open and delete "chat with us"
Just amended the sentence about port to open, and deleted the "with us". Though cloud casa is the main contirbutor for this PR, the content will be published by ovhcloud. also areduiced a bit (removed "cloud casa does all the work" as this is just a tool and does not actually take the resposnability for building and maintaining a full drp .