proxmox-network-HG-Scale: remove unnecessary sysctl parameters for vRack
Remove ip_forward and proxy_arp for the vRack section. ip_forward is not
required because there is no IP forwarding done. The VMs are in a
network (the vRack) where they communicate directly with the gateway
(the gateway answers ARP requests in the vRack), so there is no need for
proxy_arp either. I tested this in SK-1686.